requirements.txt
Flask==0.10.1
Flask-Cors==2.1.2
Flask-SQLAlchemy==2.1
flask-swagger==0.2.13
Flask-Testing==0.4.2
GeoAlchemy2==0.2.5
Jinja2==2.8
MarkupSafe==0.23
SQLAlchemy==1.0.12
Shapely==1.5.15
Werkzeug==0.11.8
factory-boy==2.5.2
fake-factory==0.5.7
itsdangerous==0.24
psycopg2==2.6.1
pyshp==1.2.3
pyslack-real==0.6.0
python-dateutil==2.5.3
requests==2.9.1
six==1.10.0
pyproj==1.9.5.1

